+++ playground/src/logics/config.ts
@@ -1,5 +1,5 @@
 import type { UserConfig } from 'unocss'
-import { createConfig } from '../../unocss.config'
+import { createConfig } from '../../unocss.config.client'
 import { options } from './url'
 
 export const defaultConfig = computed(() =>
