+++ playground/unocss.config.client.ts
@@ -1,0 +1,25 @@
+import {
+  defineConfig,
+  presetAttributify,
+  presetUno,
+  transformerCompileClass,
+  transformerDirectives,
+  transformerVariantGroup,
+} from 'unocss'
+
+export function createConfig({ strict = true, dev = true } = {}) {
+  return defineConfig({
+    envMode: dev ? 'dev' : 'build',
+    presets: [
+      presetAttributify({ strict }),
+      presetUno(),
+    ],
+    transformers: [
+      transformerVariantGroup(),
+      transformerDirectives(),
+      transformerCompileClass(),
+    ],
+  })
+}
+
+export default createConfig()
