+++ playground/unocss.config.ts
@@ -8,28 +8,21 @@ import {
   transformerVariantGroup,
 } from 'unocss'
 
-export function createConfig({ strict = true, dev = true } = {}) {
-  return defineConfig({
-    envMode: dev ? 'dev' : 'build',
-    theme: {
-      fontFamily: {
-        sans: '\'Inter\', sans-serif',
-        mono: '\'Fira Code\', monospace',
-      },
+export default defineConfig({
+  theme: {
+    fontFamily: {
+      sans: '\'Inter\', sans-serif',
+      mono: '\'Fira Code\', monospace',
     },
-    presets: [
-      presetAttributify({ strict }),
-      presetUno(),
-      presetIcons({
-        scale: 1.2,
-      }),
-    ],
-    transformers: [
-      transformerVariantGroup(),
-      transformerDirectives(),
-      transformerCompileClass(),
-    ],
-  })
-}
-
-export default createConfig()
+  },
+  presets: [
+    presetAttributify(),
+    presetUno(),
+    presetIcons(),
+  ],
+  transformers: [
+    transformerVariantGroup(),
+    transformerDirectives(),
+    transformerCompileClass(),
+  ],
+})
